Changes AnalyzedImageStructuredResponse props to decimal

+++ src/Modules/AiChat/MealMind.Modules.AiChat.Application/Dtos/DetectedFoodResponse.cs
@@ -4,13 +4,13 @@ public class DetectedFoodResponse
 {
     public string FoodName { get; init; } = string.Empty;
     public double QuantityInGrams { get; init; }
-    public double ConfidenceScore { get; init; }
-    public double MinEstimatedCalories { get; init; }
-    public double MaxEstimatedCalories { get; init; }
-    public double MinEstimatedProteins { get; init; }
-    public double MaxEstimatedProteins { get; init; }
-    public double MinEstimatedFats { get; init; }
-    public double MaxEstimatedFats { get; init; }
-    public double MinEstimatedCarbohydrates { get; init; }
-    public double MaxEstimatedCarbohydrates { get; init; }
+    public decimal ConfidenceScore { get; init; }
+    public decimal MinEstimatedCalories { get; init; }
+    public decimal MaxEstimatedCalories { get; init; }
+    public decimal MinEstimatedProteins { get; init; }
+    public decimal MaxEstimatedProteins { get; init; }
+    public decimal MinEstimatedFats { get; init; }
+    public decimal MaxEstimatedFats { get; init; }
+    public decimal MinEstimatedCarbohydrates { get; init; }
+    public decimal MaxEstimatedCarbohydrates { get; init; }
 }
